+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Splash screen for old Android versions
 
+### Fixed
+- Occasional blinking of gallery items while loading a new page
+
 ## [1.11.0] - 2023-06-18
 
 ### Added